Escape single quotes in formula field_value
I encountered a bug where Airtable responds with a 422 error when a formula value contains a single quote or apostrophe. For example, a string like "Men's Clothing" is incompatible with the single quotes used to wrap the formula. Escaping the single quote resolved the issue. This library makes my life better on a daily basis so the least I can do is contribute a (very small) improvement. Hope this is useful, thanks!