feat: Introduce RequestStrategy pattern
This commit antroduces a new pattern for consumers to inject alternate request
functionality into pyAirtable's ApiAbstract base class. A RequestStrategy should
accept the request parameters pyAirtable uses, make the HTTP request to the
Airtable API and then process the response and return the contained records in
the format that pyAirtable expects.

Three implementations are included, SimpleRequestStrategy,
RetryingRequestStrategy, and RateLimitRetryingRequestStrategy:

SimpleRequestStrategy encapsulates the use of a `requests.Session` to make the
requests, and expects the caller to provide any needed parameters like timeout
or headers. This behavior should be a 1:1 replacement of the old behavior.
SimpleRequestStrategy accepts an optional pre-configured Session if the caller
needs to do more specific set-up, e.g. of headers, etc.

RetryingRequestStrategy relies on `tenacity.Retrying`, from tenacity, a
well-established, widely trusted, and batteries-included function-retry library.
The caller must provide a configured Retrying object to control the retry
behavior on the requests.

RateLimitRetryingRequestStrategy subclasses RetryingRequestStrategy and provides
a Retrying object that reimplements the built-in retry behavior of the official
Airtable JS library, airtable.js. Requests to Airtable that hit the rate limit
of 5 request/second, are rejected with a 429 status code, as are all subsequent
requests for 30 seconds. airtable.js accommodates this with a jittered
exponential backoff starting at 30s when it receives a 429, and that behavior is
implemented here.