Unblock readthedocs build on Python 3.7
Looking at [this build failure](https://readthedocs.org/projects/pyairtable/builds/20051009/)
it appears readthedocs.org is still building on Python 3.7, which means we can't use flake8 >= 6.0.0.

It looks like the version of Python used for building documentation
[can be configured](https://docs.readthedocs.io/en/stable/config-file/v2.html)
but I'm not trying to take that on as part of this branch.

Moving the documentation build into `tox` helps make it reproducible
across different developers' environments.

I did encounter one formatter warning on the existing documentation,
which I've fixed as part of this branch.