Stop trying to type check records from the API
In 16bdcf2 one can see the sort of hoops we jump through
when we need Airtable's API responses to be type checked.
This commit gives up on that approach, since what we're
doing here is inherently not forward-compatible with any
future changes or new field types they might add.

We still type-check the arguments to `create()`/`update()`
to warn against writing values we know the API will reject.