CRUD support for comments
This branch closes #238 by adding a few new methods and one new class:
* `Api.iterate_requests(...)` is a generic "retrieve all pages" method
  suitable for various API endpoints.
* `Table.comments(record_id)` and `Model.comments()` will retrieve all
  comments for the given record.
* `Table.add_comment(record_id, text)` and `Model.add_comment(text)`
  will add a comment to the given record.
* `Comment.text = text; Comment.save()` will modify an existing comment.
* `Comment.delete()` will delete it.

This branch also fixes #230 by adding support for multiple pages to the
"get bases" metadata API. This was an easy side effect of implementing
`Api.iterate_requests()`, which we needed for comments anyways.

Lastly, this branch sets up some base classes (`AirtableModel` and
`SerializableModel`) that will be useful for other Airtable-supplied
metadata in the future, such as webhooks.