FieldSchema.save(), TableSchema.save()
This adds support for saving changes to table and field
names/descriptions. This required a significant rethink of the way we
construct URLs for `SerializableModel`. It does makea breaking change
to the `.from_api()` method, but that method does not appear to be
[in use](https://github.com/search?q=pyairtable+from_api&type=code),
so I think this is reasonable without a major version bump.